Add files via upload
Redesign in all Catalogues so each Hero is separate from their Warband, but the warband is still joined to the Hero.

Added all Armies of the Hobbit.

Duplicates in roster will now error correctly, Eg if the Witch-king is taken as part of an Angmar force as well as a Mordor force it will be recognised as the same profile. etc.